Refactor: Migrate services from TypeORM to Mongoose
This commit replaces TypeORM repository usage with Mongoose models across various services. It also fixes import paths for schemas and updates user profile update logic. Old TypeORM entities and migrations are removed.
