--- backend/src/entities/coupon.entity.ts
@@ -1,33 +1,0 @@
-import { Entity, PrimaryGeneratedColumn, Column, CreateDateColumn, UpdateDateColumn, ManyToMany, JoinTable } from 'typeorm';
-import { Course } from './course.entity';
-
-@Entity()
-export class Coupon {
-  @PrimaryGeneratedColumn('uuid')
-  id: string;
-
-  @Column({ unique: true })
-  code: string;
-
-  @Column({ nullable: true })
-  usageLimit: number;
-
-  @Column({ default: 0 })
-  usedCount: number;
-
-  @Column({ type: 'timestamp', nullable: true })
-  expiresAt: Date;
-
-  @Column({ default: true })
-  isActive: boolean;
-
-  @CreateDateColumn()
-  createdAt: Date;
-
-  @UpdateDateColumn()
-  updatedAt: Date;
-
-  @ManyToMany(() => Course)
-  @JoinTable()
-  allowedCourses: Course[];
-}
