--- backend/src/entities/course.entity.ts
@@ -1,36 +1,0 @@
-import { Entity, PrimaryGeneratedColumn, Column, CreateDateColumn, UpdateDateColumn, ManyToOne, OneToMany, Index } from 'typeorm';
-import { User } from './user.entity';
-import { Lesson } from './lesson.entity';
-
-@Entity()
-@Index(['teacherId'])
-@Index(['isActive'])
-@Index(['createdAt'])
-export class Course {
-  @PrimaryGeneratedColumn('uuid')
-  id: string;
-
-  @Column()
-  title: string;
-
-  @Column('text')
-  description: string;
-
-  @Column({ default: true })
-  isActive: boolean;
-
-  @CreateDateColumn()
-  createdAt: Date;
-
-  @UpdateDateColumn()
-  updatedAt: Date;
-
-  @ManyToOne(() => User, user => user.coursesCreated)
-  teacher: User;
-
-  @Column()
-  teacherId: string;
-
-  @OneToMany(() => Lesson, lesson => lesson.course)
-  lessons: Lesson[];
-}
