--- backend/src/entities/upload-request.entity.ts
@@ -1,44 +1,0 @@
-import { Entity, PrimaryGeneratedColumn, Column, CreateDateColumn, UpdateDateColumn, ManyToOne } from 'typeorm';
-import { User } from './user.entity';
-
-export enum UploadType {
-  VIDEO = 'video',
-  PDF = 'pdf',
-  QUIZ = 'quiz',
-}
-
-export enum UploadStatus {
-  PENDING = 'pending',
-  APPROVED = 'approved',
-  REJECTED = 'rejected',
-}
-
-@Entity()
-export class UploadRequest {
-  @PrimaryGeneratedColumn('uuid')
-  id: string;
-
-  @Column({ type: 'enum', enum: UploadType })
-  type: UploadType;
-
-  @Column({ type: 'enum', enum: UploadStatus, default: UploadStatus.PENDING })
-  status: UploadStatus;
-
-  @Column('json')
-  metadata: any;
-
-  @Column('text', { nullable: true })
-  adminNotes: string;
-
-  @CreateDateColumn()
-  createdAt: Date;
-
-  @UpdateDateColumn()
-  updatedAt: Date;
-
-  @ManyToOne(() => User, user => user.uploadRequests)
-  teacher: User;
-
-  @Column()
-  teacherId: string;
-}
