--- backend/src/entities/user.entity.ts
@@ -1,52 +1,0 @@
-import { Entity, PrimaryGeneratedColumn, Column, CreateDateColumn, UpdateDateColumn, OneToMany, ManyToMany, JoinTable, Index } from 'typeorm';
-import { Course } from './course.entity';
-import { UploadRequest } from './upload-request.entity';
-
-export enum UserRole {
-  ADMIN = 'admin',
-  TEACHER = 'teacher',
-  STUDENT = 'student',
-}
-
-@Entity()
-@Index(['email'], { unique: true })
-@Index(['role'])
-@Index(['createdAt'])
-export class User {
-  @PrimaryGeneratedColumn('uuid')
-  id: string;
-
-  @Column()
-  name: string;
-
-  @Column({ unique: true })
-  email: string;
-
-  @Column()
-  passwordHash: string;
-
-  @Column({ type: 'enum', enum: UserRole })
-  role: UserRole;
-
-  @Column({ nullable: true })
-  phone: string;
-
-  @Column({ nullable: true })
-  avatar: string;
-
-  @CreateDateColumn()
-  createdAt: Date;
-
-  @UpdateDateColumn()
-  updatedAt: Date;
-
-  @OneToMany(() => Course, course => course.teacher)
-  coursesCreated: Course[];
-
-  @ManyToMany(() => Course)
-  @JoinTable()
-  coursesUnlocked: Course[];
-
-  @OneToMany(() => UploadRequest, upload => upload.teacher)
-  uploadRequests: UploadRequest[];
-}
